refactor | Remove unused components and refactor folder structure
