--- app/(dashboard)/devices/page.tsx
@@ -1,34 +1,0 @@
-import { CardWrapper } from "@/app/components/wrappers";
-
-import { TableWithFilter, FormDevice } from "@/app/components/ui";
-
-import { headingFont} from "@/app/config/fonts";
-import {devices, detailedColumns} from "@/app/lib/data";
-import {devicesAll } from "@/app/lib/data";
-
-
-export default async function DevicePage() {
-  
-
-  return (
-   <div>
-      <h2 className={`heading-h2 ${headingFont.className}`}>Unidades</h2>
-      <CardWrapper className="mb-5">
-          <h1>Tabla</h1>
-         {/* <FormDevice data={devicesAll}/> */}
-       
-      </CardWrapper>
-      <CardWrapper>
-        <h2>Equipo</h2>
-          {/* <TableWithFilter 
-            data={devices} 
-            columns={detailedColumns}
-            showActions={true}
-            textButtonAction="Ver actividad"
-            linkHref="/device/12"
-            /> */}
-        </CardWrapper>
-
-   </div>
-  );
-}
