+++ app/components/ui/index.ts
@@ -1,8 +1,1 @@
-export * from './card/CardWithIcon';
 export * from './horizontalBar/HorizontalBar';
-export * from './forms/FormDevice';
-
-export * from './card/CardDetailDevice';
-export * from './map/MapContainer';
-export * from './header/HeaderSection'
-export * from './forms/AddDevice';
