--- app/config/firebase.ts
@@ -1,17 +1,0 @@
-import { initializeApp } from "firebase/app";
-import { getFirestore } from "firebase/firestore";
-
-const firebaseConfig = {
-  apiKey: `${process.env.NEXT_PUBLIC_FIREBASE_API_KEY}`,
-  authDomain: `${process.env.NEXT_PUBLIC_AUTH_DOMAIN}`,
-  projectId: `${process.env.NEXT_PUBLIC_PROJECT_ID}`,
-  storageBucket: `${process.env.NEXT_PUBLIC_STORAGE_BUCKET}`,
-  messagingSenderId: `${process.env.NEXT_PUBLIC_MESSAGING_SENDER_ID}`,
-  appId: `${process.env.NEXT_PUBLIC_APP_ID}`,
-};
-
-// Initialize Firebase
-const app = initializeApp(firebaseConfig);
-const db = getFirestore(app);
-export {db} 
-export default app;
