+++ app/ui/cards/cardDetails.module.css
